move copy assets to build step
