+++ packages/cli/templates/default-plugin/package.json.hbs
@@ -6,7 +6,7 @@
   "license": "Apache-2.0",
   "private": false,
   "scripts": {
-    "build": "backstage-cli plugin:copyAssets && backstage-cli plugin:build",
+    "build": "backstage-cli plugin:build",
     "lint": "backstage-cli plugin:lint",
     "test": "backstage-cli plugin:test"
   },
